backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.16",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "load-dotenv>=0.1.0",
    "pandas>=3.0.1",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.41.0",
    "yfinance>=1.2.0",
]
